Changed the field due_date from Date to Date *.
This is because if the field due_date was declared Date, then the value will be copied when it is being
retrived or when an instance of Task was created. Hence a pointer will mean that the object needs to
be created once and the address will be passed.

In addition to this both constructors were fixed for allowing the due_date field. Getter methods for
all fields have been implemented.