Testing the Task Class to see if it works as expected.
Initally there were failures, due to getting string input using cin, where the string had whitespace.
Therefore, switched to using getline() for string input.

If the zero parameter constructor was used then the due_date was set to null and within the body a
new Date object is created within the body. Also, exception handling was added if the new operator
were to fail if memory could not be allocated.

Also added destructor which will free the memory allocated for the due_date field.