Lot of changes made to make it more compaitable with class inherited from Stage
First an empty constructor was added, this is useful for helping implement polymorphism. Method insert_task(void) was removed and added only in a derived class since not all dervied classes should insert_task via stdin, which is what this method did. Methods, remove_task, modify_task, load_task, save_task and print_task all modified to be virtual as they will be overriden in derived class.

Created pure virtual method menu() was created, allowing the options to be implmented in derived class so that specific menu can be created for each derived stage. Created print method which will print a menu of all possible print methods that can be called and it will be called from this method. Created a find_id method to return the same result as find() from unordered_map. Direct all to find() is not possible as tasks field is private. Added getters for stage and tasks field. Since there are virtual methods destructor changed to be virtual as well.