Created Ongoing class
Created Ongoing class that will hold all the tasks that are currently being considered. This class inherits from the Stage Class. The class overrides the following base class methods, print_task(), load_task(), save_task() and menu().

Two other methods were created, increase_current_allocated_time() and decrease_current_allocated_time(). These were used in the trasnfer_task() function present in the main.cpp, to decrease or increase the current_allocated_time if a task was transfered to or from an instance of Ongoing.

In addition to this there are accessors that allow acces to the fields max_allocated_time and current_allocated_time.